Add files via upload
So this is the folder that contains all of the exercises within Robert LaFore's book on data structures and algorithms. Moreover, this repository contains projects of my own making. Take note of the fact that there is a C++ folder. Along with data structures and algorithms, I am learning C++ to better maintain my ability to comprehend langauges.